restructure address-input
