+++ packages/registry/components/docs/docs-account-info.tsx
@@ -57,170 +57,3 @@ export function AccountInfoDocs() {
     </PolkadotProvider>
   );
 }
-
-// ### Basic Example
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview"><AddressInputWithProvider /></TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// import { AddressInput } from "@/components/address-input.dedot";
-
-// function MyComponent() {
-//   const [address, setAddress] = useState("");
-
-//   return (
-//     <AddressInput value={address} onChange={setAddress} />
-//   );
-// }
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Custom Identity Chain
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identityChain="paseo_people"
-//       withIdentityLookup={true}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identityChain="paseo_people"  // Specify which chain to use for identity lookup
-//   withIdentityLookup={true}
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### With Identity Callback
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       withIdentityLookup={true}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   onIdentitySelected={(identity) => {
-//     console.log("Found identity:", identity.data.display);
-//   }}
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Truncated Display
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="both"
-//       truncate={6}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="both"
-//   truncate={6}  // Shows first 6 and last 6 characters
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Custom Identicon Themes
-
-// #### Beachball Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="beachball"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="beachball"  // Colorful beachball style
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// #### Substrate Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="substrate"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="substrate"  // Substrate theme
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// #### JDenticon Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="jdenticon"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="jdenticon"  // Jdenticon style
-// />
-// ```
-// </TabsContent>
-// </Tabs>
