+++ packages/registry/components/docs/docs-address-input.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/components/address-input.dedot";
-import { addressInputExample } from "../examples/dedot/example-address-input";
-import { ComponentExample } from "../examples/types.examples";
-import { ComponentPreview } from "../layout/component-preview";
+import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/address-input.dedot";
+import { addressInputExample } from "@/components/examples/dedot/example-address-input";
+import { ComponentExample } from "@/components/examples/types.examples";
+import { ComponentPreview } from "@/components/layout/component-preview";
 import { PolkadotProvider } from "@/registry/polkadot-ui/lib/polkadot-provider.dedot";
-import { IdentityResult } from "@/registry/polkadot-ui/blocks/address-input/components/address-input.base";
+import { IdentityResult } from "@/registry/polkadot-ui/blocks/address-input/address-input.base";
 
 const handleIdentitySelected = (identity: IdentityResult) => {
   alert(identity.data.display);
@@ -78,170 +78,3 @@ export function AddressInputDocs() {
     </PolkadotProvider>
   );
 }
-
-// ### Basic Example
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview"><AddressInputWithProvider /></TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// import { AddressInput } from "@/components/address-input.dedot";
-
-// function MyComponent() {
-//   const [address, setAddress] = useState("");
-
-//   return (
-//     <AddressInput value={address} onChange={setAddress} />
-//   );
-// }
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Custom Identity Chain
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identityChain="paseo_people"
-//       withIdentityLookup={true}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identityChain="paseo_people"  // Specify which chain to use for identity lookup
-//   withIdentityLookup={true}
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### With Identity Callback
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       withIdentityLookup={true}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   onIdentitySelected={(identity) => {
-//     console.log("Found identity:", identity.data.display);
-//   }}
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Truncated Display
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="both"
-//       truncate={6}
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="both"
-//   truncate={6}  // Shows first 6 and last 6 characters
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// ### Custom Identicon Themes
-
-// #### Beachball Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="beachball"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="beachball"  // Colorful beachball style
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// #### Substrate Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="substrate"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="substrate"  // Substrate theme
-// />
-// ```
-// </TabsContent>
-// </Tabs>
-
-// #### JDenticon Theme
-
-// <Tabs defaultValue="preview" className="max-w-[600px]">
-//   <TabsList>
-//     <TabsTrigger value="preview">Preview</TabsTrigger>
-//     <TabsTrigger value="code">Code</TabsTrigger>
-//   </TabsList>
-//   <TabsContent value="preview">
-//     <AddressInputWithProvider
-//       format="ss58"
-//       identiconTheme="jdenticon"
-//     />
-//   </TabsContent>
-//   <TabsContent value="code" className="my-0">
-// ```tsx
-// <AddressInput
-//   format="ss58"
-//   identiconTheme="jdenticon"  // Jdenticon style
-// />
-// ```
-// </TabsContent>
-// </Tabs>
