+++ packages/registry/registry/polkadot-ui/blocks/address-input/demo/page.dedot.tsx
@@ -1,5 +1,5 @@
 import { PolkadotProvider } from "@/registry/polkadot-ui/lib/polkadot-provider.dedot";
-import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/components/address-input.dedot";
+import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/address-input.dedot";
 import { AddressInputDemo } from "@/registry/polkadot-ui/blocks/address-input/demo/address-input-demo";
 
 export default function AddressInputDedotPage() {
