+++ packages/registry/registry/polkadot-ui/blocks/address-input/demo/page.papi.tsx
@@ -1,5 +1,5 @@
 import { PolkadotProvider } from "@/registry/polkadot-ui/lib/polkadot-provider.papi";
-import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/components/address-input.papi";
+import { AddressInput } from "@/registry/polkadot-ui/blocks/address-input/address-input.papi";
 import { AddressInputDemo } from "@/registry/polkadot-ui/blocks/address-input/demo/address-input-demo";
 
 export default function AddressInputPage() {
